Revert "Refs #93: WIP implement movement behavior"
This reverts commit ca0c0fd217a55b67229bfcf658b2c467a6c58273.

+++ servers/game/src/main/kotlin/fr/rob/game/app/node/GameNodeUpdateScheduler.kt
@@ -7,13 +7,13 @@ class GameNodeUpdateScheduler(private val instanceManager: InstanceManager) {
     fun loop() {
         var realCurrentTime: Long
         var realPreviousTime = System.currentTimeMillis()
-        var deltaTime: Int
+        var deltaTime: Long
         var executionTime: Long
 
         while (true) {
             realCurrentTime = System.currentTimeMillis()
 
-            deltaTime = (realCurrentTime - realPreviousTime).toInt()
+            deltaTime = realCurrentTime - realPreviousTime
 
             update(deltaTime)
 
@@ -26,7 +26,7 @@ class GameNodeUpdateScheduler(private val instanceManager: InstanceManager) {
         }
     }
 
-    private fun update(deltaTime: Int) {
+    private fun update(deltaTime: Long) {
         instanceManager.update(deltaTime)
     }
